+++ app/ts/background/medataSearch.ts
@@ -1,7 +1,7 @@
 import { addressString } from '../utils/bigint.js'
-import { AddressBookEntries, AddressBookEntry, ActiveAddress, ActiveAddressEntry, ContactEntry, ContractEntry, Erc1155Entry, Erc20TokenEntry, Erc721Entry } from '../types/addressBookTypes.js'
+import { AddressBookEntries, AddressBookEntry, ActiveAddress, ActiveAddressEntry, ContactEntry, ContractEntry, Erc1155Entry, Erc20TokenEntry, Erc721Entry, UserAddressBook } from '../types/addressBookTypes.js'
 import { nftMetadata, tokenMetadata, contractMetadata, NftDefinition, ContractDefinition, TokenDefinition } from '@darkflorist/address-metadata'
-import { AddressBookCategory, GetAddressBookDataFilter, UserAddressBook } from '../types/interceptor-messages.js'
+import { AddressBookCategory, GetAddressBookDataFilter } from '../types/interceptor-messages.js'
 import { getFullLogoUri } from './metadataUtils.js'
 import { assertNever } from '../utils/typescript.js'
 import { getUserAddressBookEntries } from './storageVariables.js'
