+++ app/ts/background/metadataUtils.ts
@@ -1,10 +1,9 @@
 import { addressString, checksummedAddress } from '../utils/bigint.js'
-import { ActiveAddress, ActiveAddressEntry, AddressBookEntry } from '../types/addressBookTypes.js'
+import { ActiveAddress, ActiveAddressEntry, AddressBookEntry, UserAddressBook } from '../types/addressBookTypes.js'
 import { NamedTokenId, SimulationState, VisualizerResult } from '../types/visualizer-types.js'
 import { nftMetadata, tokenMetadata, contractMetadata } from '@darkflorist/address-metadata'
 import { ethers } from 'ethers'
 import { MOCK_ADDRESS } from '../utils/constants.js'
-import { UserAddressBook } from '../types/interceptor-messages.js'
 import { EthereumClientService } from '../simulation/services/EthereumClientService.js'
 import { itentifyAddressViaOnChainInformation } from '../utils/tokenIdentification.js'
 import { assertNever } from '../utils/typescript.js'
