+++ app/ts/components/subcomponents/ChainSelector.tsx
@@ -1,6 +1,6 @@
 import { useRef, useState } from 'preact/hooks'
 import { clickOutsideAlerter } from '../ui-utils.js'
-import { RpcEntries, RpcEntry, RpcNetwork } from '../../types/visualizer-types.js'
+import { RpcEntries, RpcEntry, RpcNetwork } from '../../types/rpc.js'
 
 interface ChainSelectorParams {
 	rpcNetwork: RpcNetwork
@@ -28,7 +28,7 @@ export function ChainSelector(params: ChainSelectorParams) {
 				<p style = 'overflow: hidden; white-space: nowrap; display: block; max-width: 160px; text-overflow: ellipsis;'>{ params.rpcNetwork.name } ▼ </p>
 			</button>
 		</div>
-		<div class = 'dropdown-menu' id = 'dropdown-menu' role = 'menu' style = 'right: -10px; min-width: 160px; left: unset'>
+		<div class = 'dropdown-menu' id = 'dropdown-menu' role = 'menu' style = 'right: -10px; min-width: 15sss0px; left: unset'>
 			<div class = 'dropdown-content'>
 				{
 					params.rpcEntries.map((rpcEntry) => { return (
