+++ app/ts/components/subcomponents/coins.tsx
@@ -1,7 +1,7 @@
 import { useSignal } from '@preact/signals'
 import { getTokenAmountsWorth } from '../../simulation/priceEstimator.js'
 import { abs, bigintToDecimalString, bigintToRoundedPrettyDecimalString, checksummedAddress } from '../../utils/bigint.js'
-import { TokenPriceEstimate, RpcNetwork } from '../../types/visualizer-types.js'
+import { TokenPriceEstimate } from '../../types/visualizer-types.js'
 import { CopyToClipboard } from './CopyToClipboard.js'
 import { Blockie } from './PreactBlocky.js'
 import { JSX } from 'preact/jsx-runtime'
@@ -9,6 +9,7 @@ import { useEffect } from 'preact/hooks'
 import { Erc1155Entry, Erc20TokenEntry, Erc721Entry } from '../../types/addressBookTypes.js'
 import { RenameAddressCallBack } from '../../types/user-interface-types.js'
 import { BIG_FONT_SIZE } from '../../utils/constants.js'
+import { RpcNetwork } from '../../types/rpc.js'
 
 type EtherParams = {
 	amount: bigint
