+++ app/ts/simulation/simulator.ts
@@ -9,12 +9,13 @@ import { eoaApproval } from './protectors/eoaApproval.js'
 import { eoaCalldata } from './protectors/eoaCalldata.js'
 import { tokenToContract } from './protectors/tokenToContract.js'
 import { simulatedMulticall } from './services/SimulationModeEthereumClientService.js'
-import { WebsiteCreatedEthereumUnsignedTransaction, SimResults, SimulationState, VisualizerResult, RpcNetwork, TokenVisualizerResult } from '../types/visualizer-types.js'
+import { WebsiteCreatedEthereumUnsignedTransaction, SimResults, SimulationState, VisualizerResult, TokenVisualizerResult } from '../types/visualizer-types.js'
 import { QUARANTINE_CODE } from './protectors/quarantine-codes.js'
 import { EthereumJSONRpcRequestHandler } from './services/EthereumJSONRpcRequestHandler.js'
 import { MulticallResponse, MulticallResponseEventLog, SingleMulticallResponse } from '../types/JsonRpc-types.js'
 import { APPROVAL_LOG, DEPOSIT_LOG, ERC1155_TRANSFERBATCH_LOG, ERC1155_TRANSFERSINGLE_LOG, ERC721_APPROVAL_FOR_ALL_LOG, TRANSFER_LOG, WITHDRAWAL_LOG } from '../utils/constants.js'
 import { handleApprovalLog, handleDepositLog, handleERC1155TransferBatch, handleERC1155TransferSingle, handleERC20TransferLog, handleErc721ApprovalForAllLog, handleWithdrawalLog } from './logHandlers.js'
+import { RpcNetwork } from '../types/rpc.js'
 
 const PROTECTORS = [
 	selfTokenOops,
