+++ app/ts/utils/storageUtils.ts
@@ -1,11 +1,13 @@
 import * as funtypes from 'funtypes'
 import { EthereumAddressOrMissing } from '../types/wire-types.js'
-import { PendingAccessRequestArray, PendingChainChangeConfirmationPromise, PendingPersonalSignPromise, PendingTransaction, RpcConnectionStatus, TabState } from '../types/user-interface-types.js'
-import { EthereumSubscriptions, RpcEntries, RpcNetwork, SimulationResults } from '../types/visualizer-types.js'
+import { PendingChainChangeConfirmationPromise, PendingPersonalSignPromise, RpcConnectionStatus, TabState } from '../types/user-interface-types.js'
+import { EthereumSubscriptions, SimulationResults } from '../types/visualizer-types.js'
 import { AddressBookEntries, ActiveAddressArray, ContactEntries } from '../types/addressBookTypes.js'
 import { Page } from '../types/exportedSettingsTypes.js'
 import { WebsiteAccessArray } from '../types/websiteAccessTypes.js'
 import { SignerName } from '../types/signerTypes.js'
+import { PendingAccessRequestArray, PendingTransaction } from '../types/accessRequest.js'
+import { RpcEntries, RpcNetwork } from '../types/rpc.js'
 
 export type LocalStorageItems = funtypes.Static<typeof LocalStorageItems>
 export const LocalStorageItems = funtypes.Partial({
